population generator (random population and next generation), default board size - 8, network size in configuration, refactor

+++ app/model/ai/IArtificialInteligence.java
@@ -7,5 +7,5 @@ public interface IArtificialInteligence {
 
 	public Move makeMove(Board board);
 
-	public int evaluate(Board board);
+	public double evaluate(Board board);
 }
